requirements.txt
fastapi
uvicorn
streamlit
keras
numpy
opencv-python
Pillow
tensorflow
python-multipart
torch
torchvision
timm